fix: skip slow ops

+++ autotick-bot/install_bot_code.sh
@@ -1,26 +1,28 @@
 #!/bin/bash
 
-# clean disk space
-sudo mkdir -p /opt/empty_dir || true
-for d in \
- /opt/ghc \
- /opt/hostedtoolcache \
- /usr/lib/jvm \
- /usr/local/.ghcup \
- /usr/local/lib/android \
- /usr/local/share/powershell \
- /usr/share/dotnet \
- /usr/share/swift \
-; do
-  sudo rsync --stats -a --delete /opt/empty_dir/ $d || true
-done
-sudo apt-get purge -y -f firefox \
-                         google-chrome-stable \
-                         microsoft-edge-stable
-sudo apt-get autoremove -y >& /dev/null
-sudo apt-get autoclean -y >& /dev/null
-sudo docker image prune --all --force
-df -h
+if [[ "$1" != "--no-clean-disk-space" ]] && [[ "$2" != "--no-clean-disk-space" ]]; then
+  # clean disk space
+  sudo mkdir -p /opt/empty_dir || true
+  for d in \
+  /opt/ghc \
+  /opt/hostedtoolcache \
+  /usr/lib/jvm \
+  /usr/local/.ghcup \
+  /usr/local/lib/android \
+  /usr/local/share/powershell \
+  /usr/share/dotnet \
+  /usr/share/swift \
+  ; do
+    sudo rsync --stats -a --delete /opt/empty_dir/ $d || true
+  done
+  sudo apt-get purge -y -f firefox \
+                          google-chrome-stable \
+                          microsoft-edge-stable
+  sudo apt-get autoremove -y >& /dev/null
+  sudo apt-get autoclean -y >& /dev/null
+  sudo docker image prune --all --force
+  df -h
+fi
 
 git config --global user.name regro-cf-autotick-bot
 git config --global user.email 36490558+regro-cf-autotick-bot@users.noreply.github.com
@@ -34,10 +36,10 @@ pip install --no-deps --no-build-isolation -e .
 
 cd ..
 
-if [[ "$1" != "--no-clone-graph" ]]; then
-    git clone --depth=5 https://github.com/regro/cf-graph-countyfair.git cf-graph
+if [[ "$1" != "--no-clone-graph" ]] && [[ "$2" != "--no-clone-graph" ]]; then
+  git clone --depth=5 https://github.com/regro/cf-graph-countyfair.git cf-graph
 else
-    echo "Skipping cloning of cf-graph"
+  echo "Skipping cloning of cf-graph"
 fi
 
 bot_tag=$(python -c "import conda_forge_tick; print(conda_forge_tick.__version__)")
